realizando prueba hashset

+++ src/modelo/Producto.java
@@ -45,9 +45,12 @@ public class Producto implements Serializable {
 	public void setNombre(String nombre) {
 		this.nombre = nombre;
 	}
-	public double getValor() {
+	
+
+	public long getValor() {
 		return valor;
 	}
+
 	public void setValor(long valor) {
 		this.valor = valor;
 	}
